Funcionalidades, modificar, eliminar, y listar de la tabla Reservas agregado. Listo Para documentar.

+++ src/main/java/com/model/Reserva.java
@@ -29,11 +29,12 @@ public class Reserva {
         this.formaPago = formaPago;
     }
 
-    public Reserva(int id, String fechaEntrada, String fechaSalida, Float precio,String formaPago) {
+    public Reserva(int id, String fechaEntrada, String fechaSalida, BigDecimal precio,String formaPago) {
         this.id = id;
         this.fechaEntrada = fechaEntrada;
         this.fechaSalida = fechaSalida;
-        this.precio = BigDecimal.valueOf(precio);
+        this.precio = precio;
+//        this.precio = BigDecimal.valueOf(precio);
         this.formaPago = formaPago;
     }
